+++ sources/RAM.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 namespace libvm {
     namespace RAM {
-        uint8_t ReadIndex(uint8_t Index) {
+        uint8_t ReadIndex(long int Index) {
             if(Index == NULL /*|| Index > RAMSize*/) {DebugPrint("error while reading index, exiting..."); return 1;}
         }
     }
